메타데이터 추출 형태 변경, resource 추가, s3 컨트롤러 수정
