+++ src/main/java/com/capstone/favicon/admin/application/FAQServiceImpl.java
@@ -1,0 +1,80 @@
+package com.capstone.favicon.admin.application;
+
+import com.capstone.favicon.admin.application.service.FAQService;
+import com.capstone.favicon.admin.domain.FAQ;
+import com.capstone.favicon.admin.dto.FAQRequestDto;
+import com.capstone.favicon.admin.dto.FAQResponseDto;
+import com.capstone.favicon.admin.repository.FAQRepository;
+import com.capstone.favicon.user.domain.User;
+import com.capstone.favicon.user.repository.UserRepository;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+@RequiredArgsConstructor
+public class FAQServiceImpl implements FAQService {
+
+    private final FAQRepository faqRepository;
+    private final UserRepository userRepository;
+
+    @Override
+    public void createFAQ(Long userId, FAQRequestDto request) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new RuntimeException("사용자를 찾을 수 없습니다."));
+        FAQ faq = new FAQ();
+        faq.setUser(user);
+        faq.setCategory(request.getCategory());
+        faq.setQuestion(request.getQuestion());
+        faq.setAnswer(request.getAnswer());
+
+        faqRepository.save(faq);
+    }
+
+    @Override
+    public void updateFAQ(Long faqId, FAQRequestDto request) {
+        FAQ faq = faqRepository.findById(faqId)
+                .orElseThrow(() -> new RuntimeException("FAQ를 찾을 수 없습니다."));
+
+        faq.setCategory(request.getCategory());
+        faq.setQuestion(request.getQuestion());
+        faq.setAnswer(request.getAnswer());
+
+        faqRepository.save(faq);
+    }
+
+    @Override
+    public void deleteFAQ(Long faqId) {
+        FAQ faq = faqRepository.findById(faqId)
+                .orElseThrow(() -> new RuntimeException("FAQ를 찾을 수 없습니다."));
+        faqRepository.delete(faq);
+    }
+
+    public User getAdminUserFromSession(HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        Long id = (Long) session.getAttribute("id");
+        User user = userRepository.findByUserId(id);
+        if (user == null || user.getRole() != 1) {
+            throw new RuntimeException("이 기능은 관리자만 접근 가능합니다.");
+        }
+        return user;
+    }
+
+    @Override
+    public List<FAQResponseDto> getAllFAQs() {
+        List<FAQ> faqs = faqRepository.findAll();
+        return faqs.stream().map(FAQResponseDto::new).collect(Collectors.toList());
+    }
+
+    @Override
+    public FAQResponseDto getFAQById(Long faqId) {
+        FAQ faq = faqRepository.findById(faqId)
+                .orElseThrow(() -> new RuntimeException("FAQ를 찾을 수 없습니다."));
+        return new FAQResponseDto(faq);
+    }
+
+}
